feat(c-api) Move `instance` into its own module.
